clean up structure of model directories
